+++ Source/General/Package.vb
@@ -2863,6 +2863,8 @@ Public Class Package
                         value = x265Type.Patman
                     ElseIf output.Contains("DJATOM") Then
                         value = x265Type.DJATOM
+                    ElseIf output.Contains("JPSDR") Then
+                        value = x265Type.JPSDR
                     Else
                         value = x265Type.Vanilla
                     End If