+++ Source/General/ApplicationSettings.vb
@@ -68,7 +68,7 @@ Public Class ApplicationSettings
     Public PreviewFormBorderStyle As FormBorderStyle
     Public PreviewSize As Integer = 70
     Public ProcessPriority As ProcessPriorityClass = ProcessPriorityClass.Idle
-    Public ProgressReformatting As Boolean = False
+    Public ProgressReformatting As Boolean = True
     Public ProjectsMruNum As Integer = 10
     Public RecentFramePositions As List(Of String)
     Public RecentOptionsPage As String